third_party/WebKit/LayoutTests/http/tests/websocket/compressed-control-frame.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test whether a compressed control frame is rejected");

window.jsTestIsAsync = true;

var ws = new WebSocket("ws://127.0.0.1:8880/compressed-control-frame");
var closeEvent;

ws.onopen = function()
{
    debug("onopen() was called.");
};

ws.onmessage = function(event)
{
    var message = event.data;
    testFailed("onmessage() was called. (message = \"" + message + "\")");
};

ws.onclose = function(event)
{
    debug("onclose() was called.");
    closeEvent = event;
    shouldBeFalse("closeEvent.wasClean");
    finishJSTest();
};

ws.onerror = function(error)
{
    testPassed("onerror() was called");
};

</script>
</body>
</html>
